requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
structlog==24.1.0
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
aiokafka==0.10.0
redis[hiredis]==5.0.8
python-jose==3.3.0
grpcio==1.65.1
grpcio-tools==1.65.1
pydantic-settings==2.4.0
tenacity==8.5.0
circuitbreaker==2.0.0
prometheus-client==0.20.0
requests==2.32.3
fastavro==1.9.5
alembic==1.13.2
dependency-injector>=4.42.0
# Testing dependencies
pytest==8.3.2
pytest-asyncio==0.23.8
asynctest==0.13.0
pytest-mock==3.14.0


opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
# opentelemetry-instrumentation-asyncpg==0.47b0
# opentelemetry-instrumentation-requests==0.47b0
# opentelemetry-instrumentation-redis==0.47b0
# opentelemetry-instrumentation-grpc==0.47b0
# opentelemetry-semantic-conventions==0.47b0





# aiokafka==0.11.0
# alembic==1.13.3
# annotated-types==0.7.0
# anyio==4.9.0
# asgiref==3.8.1
# async-timeout==5.0.1
# bcrypt==4.3.0
# certifi==2025.4.26
# cffi==1.17.1
# charset-normalizer==3.4.2
# click==8.2.1
# colorama==0.4.6
# cryptography==45.0.3
# Deprecated==1.2.18
# ecdsa==0.19.1
# fastapi==0.115.0
# fastavro==1.9.7
# googleapis-common-protos==1.56.1
# greenlet==3.2.3
# grpcio==1.73.0
# grpcio-tools==1.66.2
# h11==0.16.0
# idna==3.10
# importlib_metadata==8.4.0
# Mako==1.3.10
# MarkupSafe==3.0.2
# opentelemetry-api==1.27.0
# opentelemetry-exporter-jaeger==1.21.0
# opentelemetry-exporter-jaeger-proto-grpc==1.21.0
# opentelemetry-exporter-jaeger-thrift==1.21.0
# opentelemetry-instrumentation==0.48b0
# opentelemetry-instrumentation-asgi==0.48b0
# opentelemetry-instrumentation-fastapi==0.48b0
# opentelemetry-sdk==1.27.0
# opentelemetry-semantic-conventions==0.48b0
# opentelemetry-util-http==0.48b0
# packaging==25.0
# passlib==1.7.4
# prometheus_client==0.21.0
# protobuf==5.29.5
# pyasn1==0.6.1
# pycparser==2.22
# pydantic==2.9.2
# pydantic-settings==2.9.1
# pydantic_core==2.23.4
# python-dotenv==1.1.0
# python-jose==3.3.0
# python-json-logger==2.0.7
# redis==5.0.8
# requests==2.32.3
# rsa==4.9.1
# setuptools==80.9.0
# six==1.17.0
# sniffio==1.3.1
# SQLAlchemy==2.0.35
# starlette==0.38.6
# tenacity==9.1.2
# thrift==0.22.0
# typing-inspection==0.4.1
# typing_extensions==4.14.0
# urllib3==2.4.0
# uvicorn==0.30.6
# wrapt==1.17.2
# zipp==3.22.0
